Move the Grape middleware to the rack directory
Move the Appsignal::Grape::Middleware constant to
Appsignal::Rack::GrapeMiddleware so it matches the format of some of the
other middleware we have.

Move it to the `rack/` directory as well so all the middleware are in
one place. People can still require the `appsignal/integrations/grape`
file like previously, same as we support for
`appsignal/integrations/sinatra`. The integrations file can stay to
automate or include any other things we need.
